+++ packages/jacaranda/src/starters/index.js
@@ -1,4 +1,5 @@
 export { default as startWorker } from './worker';
 export { default as startLoopWorker } from './loopWorker';
 export { default as startCommand } from './command';
+export { default as startScheduler, cronPatterns } from './schedulerWorker';